requirements.txt
yfinance>=0.2.31
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0
streamlit>=1.30.0
plotly>=5.18.0
alpaca-py>=0.21.0
